+++ packages/logger/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./types";
-export { default as Logger } from "./Logger";
+export * from "./types/index";
+export * from "./Logger";