+++ packages/nested-collections/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./Collection";
+export * from "./Events";